backend/gee_queries.py
"""Google Earth Engine queries for satellite data analysis"""
import random
from utils import (
    get_state_boundary,
    normalize_irradiance,
    normalize_slope,
    score_land_cover,
    land_cover_name,
    estimate_grid_distance_km,
    estimate_grid_score,
    generate_sample_coordinates
)
from real_locations import get_real_locations_for_state

# NOTE: This is a simplified version for the MVP that generates realistic data
# In production, you would use actual Google Earth Engine API calls

def analyze_solar_sites(region_name, num_sites=10, criteria_weights=None, constraints=None):
    """
    Analyze potential solar sites in a given region

    For MVP: Uses real-world coordinates of suitable locations
    For Production: Would use real ee.ImageCollection queries

    Args:
        region_name: State name (e.g., "Arizona")
        num_sites: Number of top sites to return
        criteria_weights: Dict with irradiance, slope, grid_distance, land_cover
        constraints: Dict with max_slope, min_acreage, etc.

    Returns:
        List of top-ranked sites with scores and metrics
    """

    # Default weights
    if not criteria_weights:
        criteria_weights = {
            'irradiance': 0.4,
            'slope': 0.3,
            'grid_distance': 0.2,
            'land_cover': 0.1
        }

    # Default constraints
    if not constraints:
        constraints = {}

    # Try to get real locations first
    real_locations = get_real_locations_for_state(region_name, num_sites * 3)

    if real_locations:
        # Use real-world coordinates with actual location names
        sample_points = [(loc['lat'], loc['lon'], loc['name'], loc['type'])
                        for loc in real_locations]
    else:
        # Fallback to random generation for states not in database
        boundary = get_state_boundary(region_name)
        coords = generate_sample_coordinates(boundary, num_points=100, state_name=region_name)
        sample_points = [(lat, lon, None, None) for lat, lon in coords]

    # Score each point
    scored_sites = []

    for point_data in sample_points:
        if len(point_data) == 4:
            lat, lon, location_name, location_type = point_data
        else:
            lat, lon = point_data
            location_name, location_type = None, None
        # Simulate GEE data (in production, these would be actual values from datasets)
        irradiance = simulate_solar_irradiance(lat, lon, region_name)
        slope = simulate_slope(lat, lon)
        land_cover_code = simulate_land_cover(lat, lon)
        grid_distance = estimate_grid_distance_km()

        # Calculate individual scores
        irr_score = normalize_irradiance(irradiance)
        slope_score = normalize_slope(slope)
        lc_score = score_land_cover(land_cover_code)
        grid_score = estimate_grid_score(grid_distance)

        # Weighted final score
        final_score = (
            irr_score * criteria_weights['irradiance'] +
            slope_score * criteria_weights['slope'] +
            grid_score * criteria_weights['grid_distance'] +
            lc_score * criteria_weights['land_cover']
        )

        site_data = {
            'coordinates': {'lat': lat, 'lon': lon},
            'score': round(final_score, 1),
            'metrics': {
                'solar_irradiance': round(irradiance, 2),
                'solar_irradiance_score': round(irr_score, 0),
                'slope': round(slope, 2),
                'slope_score': round(slope_score, 0),
                'grid_distance': grid_distance,
                'grid_distance_score': round(grid_score, 0),
                'land_cover': land_cover_name(land_cover_code),
                'land_cover_score': round(lc_score, 0)
            }
        }

        # Add location information if available
        if location_name:
            site_data['location_name'] = location_name
            site_data['location_type'] = location_type

        scored_sites.append(site_data)

    # Apply hard constraints (filter out sites that don't meet requirements)
    filtered_sites = scored_sites

    # Filter by max slope if specified
    max_slope = constraints.get('max_slope')
    if max_slope is not None:
        filtered_sites = [site for site in filtered_sites if site['metrics']['slope'] <= max_slope]
        print(f"[GEE] Filtered to {len(filtered_sites)} sites with slope <= {max_slope}°")

    # Filter by minimum acreage if specified (currently we don't track individual site acreage,
    # but we could filter based on having suitable conditions for that acreage)
    min_acreage = constraints.get('min_acreage')
    if min_acreage:
        # For larger acreage, favor sites with better overall scores
        # This is a simplified approach - in production you'd check actual land availability
        if min_acreage > 100:
            # For large sites, require higher scores
            filtered_sites = [site for site in filtered_sites if site['score'] >= 70]
            print(f"[GEE] Filtered to {len(filtered_sites)} sites suitable for {min_acreage}+ acres")

    # Sort by score and return top N
    filtered_sites.sort(key=lambda x: x['score'], reverse=True)

    # Add rank to each site
    for idx, site in enumerate(filtered_sites[:num_sites]):
        site['rank'] = idx + 1

    return filtered_sites[:num_sites]

def simulate_solar_irradiance(lat, lon, region):
    """
    Simulate solar irradiance based on latitude and region
    Based on NASA POWER GHI dataset patterns
    """
    # Base irradiance varies by latitude (closer to equator = higher)
    base = 5.5 - (abs(lat - 25) * 0.05)

    # Regional modifiers based on actual US solar patterns
    regional_bonus = {
        'Arizona': 1.3,
        'California': 1.1,
        'Nevada': 1.2,
        'Texas': 0.8,
        'New Mexico': 1.2,
        'Utah': 1.0,
        'Colorado': 0.9,
        'Oklahoma': 0.7,
        'Kansas': 0.6,
        'Florida': 0.9,
    }.get(region, 0.5)

    # Add some randomness
    noise = random.uniform(-0.3, 0.3)

    return max(3.0, min(7.5, base + regional_bonus + noise))

def simulate_slope(lat, lon):
    """
    Simulate terrain slope
    Based on SRTM elevation data patterns
    """
    # Most sites should be relatively flat (good for solar)
    # Use exponential distribution to favor lower slopes
    slope = abs(random.gauss(3, 4))
    return min(15.0, slope)

def simulate_land_cover(lat, lon):
    """
    Simulate land cover type
    Based on ESA WorldCover classifications
    """
    # Weight towards suitable land types
    suitable_types = [60, 30, 40]  # barren, grassland, cropland
    unsuitable_types = [10, 20, 50, 80]  # forest, shrubland, built-up, water

    # 70% chance of suitable land cover
    if random.random() < 0.7:
        return random.choice(suitable_types)
    else:
        return random.choice(unsuitable_types)

# Real GEE implementation (commented out for MVP - uncomment when GEE is set up)
"""
import ee

def analyze_solar_sites_real(region_name, num_sites=10, criteria_weights=None):
    '''
    Real implementation using Google Earth Engine API
    '''
    if not criteria_weights:
        criteria_weights = {
            'irradiance': 0.4,
            'slope': 0.3,
            'grid_distance': 0.2,
            'land_cover': 0.1
        }

    # Initialize Earth Engine (should be done once in main app)
    # ee.Initialize()

    # Define region of interest
    min_lon, min_lat, max_lon, max_lat = get_state_boundary(region_name)
    roi = ee.Geometry.Rectangle([min_lon, min_lat, max_lon, max_lat])

    # Get Solar Irradiance from NASA POWER
    irradiance = ee.ImageCollection('NASA/POWER/Global_Horizontal_Irradiance') \\
        .filterBounds(roi) \\
        .select('GHI') \\
        .mean()

    # Get Slope from SRTM Elevation
    elevation = ee.Image('USGS/SRTMGL1_003')
    slope = ee.Terrain.slope(elevation)

    # Get Land Cover from ESA WorldCover
    land_cover = ee.Image('ESA/WorldCover/v200/2021').select('Map')

    # Combine all bands
    composite = irradiance.addBands(slope).addBands(land_cover)

    # Sample random points across region
    points = composite.sample(
        region=roi,
        scale=1000,  # 1km resolution
        numPixels=2000,
        geometries=True
    )

    # Convert to Python list
    points_list = points.getInfo()['features']

    # Score each point (same logic as simulate version)
    # ... rest of implementation

    return scored_sites[:num_sites]
"""
